# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
  /api/categories:
    get:
      summary: 'Get Categories'
      operationId: getCategories
      description: 'Display a listing of Categories.'
      parameters:
        -
          in: query
          name: page
          description: 'Which page to show.'
          example: 12
          required: false
          schema:
            type: integer
            description: 'Which page to show.'
            example: 12
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: 1
                      name: deleniti
                    -
                      id: 2
                      name: ipsum
                    -
                      id: 3
                      name: quis
                    -
                      id: 4
                      name: ullam
                    -
                      id: 5
                      name: quas
                    -
                      id: 6
                      name: fugiat
                    -
                      id: 7
                      name: minima
                    -
                      id: 8
                      name: est
                    -
                      id: 9
                      name: quo
                    -
                      id: 10
                      name: culpa
                    -
                      id: 11
                      name: 'new category'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: 1
                        name: deleniti
                      -
                        id: 2
                        name: ipsum
                      -
                        id: 3
                        name: quis
                      -
                        id: 4
                        name: ullam
                      -
                        id: 5
                        name: quas
                      -
                        id: 6
                        name: fugiat
                      -
                        id: 7
                        name: minima
                      -
                        id: 8
                        name: est
                      -
                        id: 9
                        name: quo
                      -
                        id: 10
                        name: culpa
                      -
                        id: 11
                        name: 'new category'
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: deleniti
      tags:
        - Categories
      security: []
    post:
      summary: 'POST Category'
      operationId: pOSTCategory
      description: 'Store a newly created category in storage.'
      parameters: []
      responses: {  }
      tags:
        - Categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: 'Name of the category.'
                  example: '"Clothing"'
              required:
                - name
      security: []
...